backend/requirements.txt
# MemoryBook Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0

# Data Validation
pydantic==2.5.3

# Async Support
aiohttp==3.9.1
aiofiles==23.2.1

# Image Processing
Pillow==10.2.0

# File Upload Support
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Google GenAI SDK (nova biblioteca oficial para Gemini)
google-genai>=1.0.0

# Firebase Admin SDK
firebase-admin>=6.5.0

# Environment variables
python-dotenv>=1.0.0

# Optional - LangGraph for validation loops
# langgraph==0.0.20
